Add KMC power strip support
There are two varieties, both have 4 switched outlets and a switched
bank of USB ports.  One has an always-on outlet and 3 USB ports (1 w/
QC) versus the other with 4 USB ports, no always-on outlet.  The
latter seems to be an older model, mine actually contains an ESP-12F
while the other uses the regular TYWE3S.  None of the extra pins on
the ESP-12F are used and I didn't bother to attempt to take advantage
of the extra flash this part might have.  Works the same as the
previous support for the 4-outlet plug, except there's no power meter
in these.  Also switched back so that single outlet devices don't
require a range.

Note that the 4-outlet + always-on + USB-QC does not have an LED to
indicate the USB power state.  There is an unpopulated LED and
resistor spot on the board, should you have some available to install.
For the other model, the LED is driven from the USB power supply and
had some serious lag in turning on and off.  The rate at which it
turns off depends on the load on the USB charger.

Signed-off-by: Alex Williamson <alex.l.williamson@gmail.com>